// api/github-sls-rest-api/api.ts
export const GithubSlsRestApiTotpMethod = {
    App: 'app',
    Email: 'email'
} as const;

export type GithubSlsRestApiTotpMethod = typeof GithubSlsRestApiTotpMethod[keyof typeof GithubSlsRestApiTotpMethod];


/**
 * 
 * @export
 * @interface GithubSlsRestApiTotpRequest
 */
export interface GithubSlsRestApiTotpRequest {
    /**
     * 
     * @type {string}
     * @memberof GithubSlsRestApiTotpRequest
     */
    'invitation': string;
}
// ...
export class TotpApi extends BaseAPI {
    /**
     * 
     * @param {string} org 
     * @param {GithubSlsRestApiTotpMethod} method 
     * @param {GithubSlsRestApiTotpRequest} githubSlsRestApiTotpRequest 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof TotpApi
     */
    public totpEnroll(org: string, method: GithubSlsRestApiTotpMethod, githubSlsRestApiTotpRequest: GithubSlsRestApiTotpRequest, options?: AxiosRequestConfig) {
        return TotpApiFp(this.configuration).totpEnroll(org, method, githubSlsRestApiTotpRequest, options).then((request) => request(this.axios, this.basePath));
    }
}